[unique] Make collectable tracking a bit more memory efficient

+++ src/game/bingo_tracking_collectables.h
@@ -5,7 +5,7 @@
 #include "area.h"
 #include "bingo.h"
 
-u16 get_unique_id(enum BingoObjectiveUpdate, f32 posX, f32 posY, f32 posZ);
-s32 is_new_kill(enum BingoObjectiveUpdate type, u16 uid);
+u32 get_unique_id(enum BingoObjectiveUpdate, f32 posX, f32 posY, f32 posZ);
+s32 is_new_kill(enum BingoObjectiveUpdate type, u32 uid);
 
 #endif /* _BINGO_TRACKING_COLLECTABLES */
